clean up attribute names
