+++ Sources/Honeycomb/UIKit/WindowInstrumentation.swift
@@ -3,70 +3,7 @@ import OpenTelemetryApi
 import SwiftUI
 import UIKit
 
-struct ViewNames {
-    var accessibilityLabel: String?
-    var accessibilityIdentifier: String?
-    var currentTitle: String?
-    var titleLabelText: String?
-    var className: String?
-
-    var name: String? {
-        return accessibilityIdentifier ?? accessibilityLabel ?? currentTitle ?? titleLabelText
-    }
-
-    init(view: UIView) {
-        findNames(view: view)
-    }
-
-    private mutating func findNames(view: UIView) {
-        // Gather various identifiers about the view.
-        if let identifier = view.accessibilityIdentifier {
-            self.accessibilityIdentifier = identifier
-        }
-        if let label = view.accessibilityLabel {
-            self.accessibilityLabel = label
-        }
-        if let button = view as? UIButton {
-            if let title = button.currentTitle {
-                self.currentTitle = title
-            }
-            if let label = button.titleLabel?.text {
-                self.titleLabelText = label
-            }
-        }
-
-        // If we've gotten _some_ identifier, stop. Otherwise, walk up the hierarchy.
-        if self.name == nil {
-            if let parent = view.superview {
-                self.findNames(view: parent)
-            }
-        }
-
-        // Set the class name for the bottom-most view.
-        self.className = String(describing: type(of: view))
-    }
-
-    func setAttributes(span: Span) {
-        if let accessibilityLabel = self.accessibilityLabel {
-            span.setAttribute(key: "view.accessibilityLabel", value: accessibilityLabel)
-        }
-        if let accessibilityIdentifier = self.accessibilityIdentifier {
-            span.setAttribute(key: "view.accessibilityIdentifier", value: accessibilityIdentifier)
-        }
-        if let currentTitle = self.currentTitle {
-            span.setAttribute(key: "view.currentTitle", value: currentTitle)
-        }
-        if let titleLabelText = self.titleLabelText {
-            span.setAttribute(key: "view.titleLabel.text", value: titleLabelText)
-        }
-        if let name = self.name {
-            span.setAttribute(key: "view.name", value: name)
-        }
-        if let className = self.className {
-            span.setAttribute(key: "view.class", value: className)
-        }
-    }
-}
+// A helper for getting attributes about a UIView that was interacted with.
 
 enum TouchType {
     case began
@@ -83,14 +20,14 @@ private func recordTouch(_ touch: UITouch, type: TouchType) {
         }
 
     // Try to find the name of the view this touch was on.
-    let viewNames: ViewNames? = touch.view.map({ view in ViewNames(view: view) })
+    let viewAttrs: ViewAttributes? = touch.view.map({ view in ViewAttributes(view: view) })
 
     let tracer = OpenTelemetry.instance.tracerProvider.get(
         instrumentationName: honeycombUIKitInstrumentationName,
         instrumentationVersion: honeycombLibraryVersion
     )
     let span = tracer.spanBuilder(spanName: spanName).startSpan()
-    viewNames?.setAttributes(span: span)
+    viewAttrs?.setAttributes(span: span)
     span.end()
 
     // Do a special check for button clicks.
@@ -99,7 +36,7 @@ private func recordTouch(_ touch: UITouch, type: TouchType) {
             if button.isHighlighted {
                 let span = tracer.spanBuilder(spanName: "click")
                     .startSpan()
-                viewNames?.setAttributes(span: span)
+                viewAttrs?.setAttributes(span: span)
                 span.end()
             }
         }
